# Conflicts:
#	modules/runner/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItDataTypesTest.java

+++ modules/runner/src/integrationTest/java/org/apache/ignite/internal/sql/engine/ItSortAggregateTest.java
@@ -118,10 +118,10 @@ public class ItSortAggregateTest extends AbstractBasicIntegrationTest {
         assertEquals(ROWS / 10, res.size());
 
         res.forEach(r -> {
-            Integer s0 = (Integer) r.get(0);
-            Integer s1 = (Integer) r.get(1);
+            long s0 = (Long) r.get(0);
+            long s1 = (Long) r.get(1);
 
-            assertEquals(s0 * 2, (int) s1);
+            assertEquals(s0 * 2, s1);
         });
     }
